Wrong variable name and case sensitive filenames
The variabel namespace defined in line 9 is unused. You use a different variable name in the script ( lines 24, 30, 35, 38), which is always an empty string, because it is never defined before.

Additionally you refer to a file uid1000.json, which does not exist on a case-sensitive filesystem, as the filename is UID1000.json.